requirements.txt
altgraph==0.17.5
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
bcrypt==4.0.1
brotli==1.2.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
cssselect2==0.8.0
docxtpl==0.20.2
et-xmlfile==2.0.0
exceptiongroup==1.3.0
fastapi==0.121.1
fonttools==4.60.1
h11==0.16.0
idna==3.11
importlib-metadata==8.7.1
jinja2==3.1.6
lxml==6.0.2
markupsafe==3.0.3
openpyxl==3.1.5
packaging==26.0
passlib==1.7.4
pefile==2024.8.26
pillow==11.3.0
psycopg2-binary==2.9.11
pycparser==2.23
pydantic==2.12.4
pydantic-core==2.41.5
pydyf==0.11.0
pyinstaller==6.18.0
pyinstaller-hooks-contrib==2026.0
PyJWT==2.10.1
pyphen==0.17.2
python-docx==1.2.0
python-dotenv==1.2.1
python-multipart==0.0.20
pywin32-ctypes==0.2.3
requests==2.32.5
sniffio==1.3.1
starlette==0.49.3
tinycss2==1.4.0
tinyhtml5==2.0.0
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.6.3
uvicorn==0.38.0
weasyprint==66.0
webencodings==0.5.1
zipp==3.23.0
zopfli==0.2.3.post1
